+++ src/engine/audio/private/voice_gain.cpp
@@ -1,6 +1,6 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
-// Purpose: 
+// Purpose:
 //
 // $NoKeywords: $
 //=============================================================================//
@@ -24,7 +24,7 @@ void CAutoGain::Reset(int blockSize, float maxGain, float avgToMaxVal, float sca
 	m_BlockSize = blockSize;
 	m_MaxGain = maxGain;
 	m_AvgToMaxVal = avgToMaxVal;
-	
+
 	m_CurBlockOffset = 0;
 	m_CurTotal = 0;
 	m_CurMax = 0;
@@ -33,7 +33,7 @@ void CAutoGain::Reset(int blockSize, float maxGain, float avgToMaxVal, float sca
 	m_NextGain = 1;
 
 	m_Scale = scale;
-	
+
 	m_GainMultiplier = 0;
 	m_FixedCurrentGain = 1 << AG_FIX_SHIFT;
 }
@@ -54,19 +54,19 @@ void CAutoGain::ProcessSamples(
 		{
 			// Update running totals..
 			m_CurTotal += abs( pCurPos[iSample] );
-			m_CurMax = max( m_CurMax, abs( pCurPos[iSample] ) );
-			
+			m_CurMax = max( m_CurMax, (int)abs( pCurPos[iSample] ) );
+
 			// Apply gain on this sample.
 			AGFixed gain = m_FixedCurrentGain + m_CurBlockOffset * m_GainMultiplier;
 			m_CurBlockOffset++;
-			
+
 			int newval = ((int)pCurPos[iSample] * gain) >> AG_FIX_SHIFT;
 			newval = min(32767, max(newval, -32768));
 			pCurPos[iSample] = (short)newval;
 		}
 		pCurPos += nToProcess;
 		nSamplesLeft -= nToProcess;
-	
+
 		// Did we just end a block? Update our next gain.
 		if((m_CurBlockOffset % m_BlockSize) == 0)
 		{
@@ -79,7 +79,7 @@ void CAutoGain::ProcessSamples(
 			float modifiedMax = avg + (m_CurMax - avg) * m_AvgToMaxVal;
 			m_NextGain = min(32767.0f / modifiedMax, m_MaxGain) * m_Scale;
 
-			// Setup the interpolation multiplier.			
+			// Setup the interpolation multiplier.
 			float fGainMultiplier = (m_NextGain - m_CurrentGain) / (m_BlockSize - 1);
 			m_GainMultiplier = (AGFixed)((double)fGainMultiplier * (1 << AG_FIX_SHIFT));
