+++ src/engine/clockdriftmgr.cpp
@@ -1,4 +1,4 @@
-//========= Copyright © 1996-2005, Valve Corporation, All rights reserved. ============//
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -180,8 +180,8 @@ void CClockDriftMgr::ShowDebugInfo( float flAdjustment )
 		int exactDiff = cl.GetClientTickCount() - m_nServerTick;
 		for ( int i=0; i < NUM_CLOCKDRIFT_SAMPLES; i++ )
 		{
-			high = max( high, m_ClockOffsets[i] );
-			low = min( low, m_ClockOffsets[i] );
+			high = max( (float)high, m_ClockOffsets[i] );
+			low = min( (float)low, m_ClockOffsets[i] );
 		}
 
 		Msg( "Clock drift: adjustment (per sec): %.2fms, avg: %.3f, lo: %d, hi: %d, ex: %d\n", flAdjustment*1000.0f, GetCurrentClockDifference(), low, high, exactDiff );
