+++ src/materialsystem/shaderapidx9/wmi.h
@@ -1,4 +1,4 @@
-//====== Copyright © 1996-2005, Valve Corporation, All rights reserved. =======
+//========= Copyright Valve Corporation, All rights reserved. ============//
 //
 // Purpose: 
 //
@@ -11,7 +11,7 @@
 #endif
 
 
-int GetVidMemBytes( void );
+uint64 GetVidMemBytes( void );
 
 
 #endif // WMI_H
